cloudup: move node-local provider code into leaf metadata packages
Extract the bootstrap authenticators and the metadata/IMDS/config helpers
that run on nodes into per-provider leaf packages (azuremetadata,
dometadata, hetznermetadata, linodemetadata, openstackmetadata,
openstackcloudconfig, scalewaymetadata), following the existing
gce/gcemetadata and gce/tpm/gcetpmsigner pattern. nodeup and protokube now
link only these leaves instead of each provider's full cloud
implementation; the openstack package in particular pulled in the full
Kubernetes clientset.

This cuts the stripped linux/arm64 nodeup binary from 122MB to 108MB and
reduces the linked k8s.io/api group/version packages from 55 to just
core/v1.
