+++ nodeup/pkg/model/kubelet.go
@@ -45,7 +45,7 @@ import (
 	"k8s.io/kops/pkg/rbac"
 	"k8s.io/kops/pkg/systemd"
 	"k8s.io/kops/upup/pkg/fi"
-	azurecloud "k8s.io/kops/upup/pkg/fi/cloudup/azure"
+	"k8s.io/kops/upup/pkg/fi/cloudup/azure/azuremetadata"
 	"k8s.io/kops/upup/pkg/fi/nodeup/nodetasks"
 	"k8s.io/kops/util/pkg/distributions"
 	kubeletv1 "k8s.io/kubelet/config/v1"
@@ -126,7 +126,7 @@ func (b *KubeletBuilder) Build(c *fi.NodeupModelBuilderContext) error {
 			}
 			providerID = fmt.Sprintf("aws:///%s/%s", instanceIdentity.AvailabilityZone, instanceIdentity.InstanceID)
 		} else if b.CloudProvider() == kops.CloudProviderAzure {
-			metadata, err := azurecloud.QueryComputeInstanceMetadata(ctx)
+			metadata, err := azuremetadata.QueryComputeInstanceMetadata(ctx)
 			if err != nil {
 				return fmt.Errorf("error querying Azure instance metadata: %v", err)
 			}
