+++ upup/pkg/fi/cloudup/hetzner/hetznermetadata/authenticator.go
@@ -14,7 +14,10 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package hetzner
+// Package hetznermetadata provides the node-local (metadata-service based)
+// parts of the Hetzner Cloud support, kept separate so that nodeup does not
+// link the full cloud provider implementation.
+package hetznermetadata
 
 import (
 	"fmt"
