+++ upup/pkg/fi/cloudup/openstack/openstackmetadata/authenticator.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package openstack
+package openstackmetadata
 
 import (
 	"fmt"
