+++ upup/pkg/fi/cloudup/openstack/openstackmetadata/testdata/metadata_drive.json
