+++ upup/pkg/fi/cloudup/openstack/openstackmetadata/testdata/metadata_service.json
