+++ upup/pkg/fi/cloudup/scaleway/scalewaymetadata/profile_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package scaleway
+package scalewaymetadata
 
 import (
 	"errors"
